requirements.txt
Flask==2.2.5
Flask-WTF==1.2.1
Werkzeug==2.3.7
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==0.21.0
SQLAlchemy==1.4.48
gunicorn==21.2.0
waitress==2.1.1
email_validator==1.3.0
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.7
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
redis==4.6.0